+++ opengl/opengl_test.go
@@ -5,6 +5,7 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 
 	"github.com/jacekolszak/pixiq/image"
 	"github.com/jacekolszak/pixiq/opengl"
@@ -169,5 +170,134 @@ func TestRun(t *testing.T) {
 			})
 		})
 		assert.NotNil(t, actualGL)
+	})
+}
+
+func TestOpenGL_Open(t *testing.T) {
+	t.Run("should constrain width to platform-specific minimum if negative", func(t *testing.T) {
+		openGL := opengl.New(mainThreadLoop)
+		defer openGL.Destroy()
+		// when
+		win := openGL.Open(-1, 0)
+		defer win.Close()
+		// then
+		require.NotNil(t, win)
+		assert.GreaterOrEqual(t, win.Width(), 0)
+	})
+	t.Run("should constrain height to platform-specific minimum if negative", func(t *testing.T) {
+		openGL := opengl.New(mainThreadLoop)
+		defer openGL.Destroy()
+		// when
+		win := openGL.Open(0, -1)
+		defer win.Close()
+		// then
+		require.NotNil(t, win)
+		assert.GreaterOrEqual(t, win.Height(), 0)
+	})
+	t.Run("should open Window", func(t *testing.T) {
+		openGL := opengl.New(mainThreadLoop)
+		defer openGL.Destroy()
+		// when
+		win := openGL.Open(640, 360)
+		defer win.Close()
+		// then
+		require.NotNil(t, win)
+		assert.Equal(t, 640, win.Width())
+		assert.Equal(t, 360, win.Height())
+	})
+	t.Run("should open two windows at the same time", func(t *testing.T) {
+		openGL := opengl.New(mainThreadLoop)
+		defer openGL.Destroy()
+		// when
+		win1 := openGL.Open(640, 360)
+		defer win1.Close()
+		win2 := openGL.Open(320, 180)
+		defer win2.Close()
+		// then
+		require.NotNil(t, win1)
+		assert.Equal(t, 640, win1.Width())
+		assert.Equal(t, 360, win1.Height())
+		require.NotNil(t, win2)
+		assert.Equal(t, 320, win2.Width())
+		assert.Equal(t, 180, win2.Height())
+	})
+	t.Run("should open another Window after first one was closed", func(t *testing.T) {
+		openGL := opengl.New(mainThreadLoop)
+		defer openGL.Destroy()
+		win1 := openGL.Open(640, 360)
+		win1.Close()
+		// when
+		win2 := openGL.Open(320, 180)
+		defer win2.Close()
+		// then
+		require.NotNil(t, win2)
+		assert.Equal(t, 320, win2.Width())
+		assert.Equal(t, 180, win2.Height())
+	})
+	t.Run("should skip nil option", func(t *testing.T) {
+		openGL := opengl.New(mainThreadLoop)
+		defer openGL.Destroy()
+		// when
+		win := openGL.Open(0, 0, nil)
+		defer win.Close()
+	})
+	t.Run("zoom <= 1 should not affect the width and height", func(t *testing.T) {
+		tests := map[string]struct {
+			zoom int
+		}{
+			"zoom = -1": {
+				zoom: -1,
+			},
+			"zoom = 0": {
+				zoom: 0,
+			},
+			"zoom = 1": {
+				zoom: 1,
+			},
+		}
+		for name, test := range tests {
+			t.Run(name, func(t *testing.T) {
+				openGL := opengl.New(mainThreadLoop)
+				defer openGL.Destroy()
+				// when
+				win := openGL.Open(640, 360, opengl.Zoom(test.zoom))
+				defer win.Close()
+				// then
+				require.NotNil(t, win)
+				assert.Equal(t, 640, win.Width())
+				assert.Equal(t, 360, win.Height())
+			})
+		}
+	})
+	t.Run("zoom should affect the width and height", func(t *testing.T) {
+		tests := map[string]struct {
+			zoom           int
+			expectedWidth  int
+			expectedHeight int
+		}{
+			"zoom = 2": {
+				zoom:           2,
+				expectedWidth:  1280,
+				expectedHeight: 720,
+			},
+			"zoom = 3": {
+				zoom:           3,
+				expectedWidth:  1920,
+				expectedHeight: 1080,
+			},
+		}
+		for name, test := range tests {
+			t.Run(name, func(t *testing.T) {
+				openGL := opengl.New(mainThreadLoop)
+				defer openGL.Destroy()
+				// when
+				win := openGL.Open(640, 360, opengl.Zoom(test.zoom))
+				defer win.Close()
+				// then
+				require.NotNil(t, win)
+				assert.Equal(t, test.expectedWidth, win.Width())
+				assert.Equal(t, test.expectedHeight, win.Height())
+			})
+		}
 	})
 }
