+++ opengl/window_bench_test.go
